// tests/smoke/mt-api.schema.ts
export const PostV1OrdersResponse = z.object({
    "orderId": z.number(),
    "code": z.string(),
    "message": z.string()
})
// ...
export const GetV1PingResponse = z.object({
    "status": z.string().optional()
})